Correção de erros no ContaController

+++ src/main/java/br/com/cdb/dao/impl/ContaDaoImpl.java
@@ -48,7 +48,7 @@ public class ContaDaoImpl implements ContaDao {
 	}
 
 	@Override
-	public boolean transferenciaTed(long agencia, int numeroConta, double valor, TipoPagamento tipo, long id) {
+	public boolean transferenciaTed(int agencia, int numeroConta, double valor, TipoPagamento tipo, long id) {
 
 		Conta contaD = new Conta();
 		Conta contaO = new Conta();
